refactor: rename native select, option
